+++ src/App.tsx
@@ -13,8 +13,8 @@ import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import React, { createContext, useContext, useState } from 'react';
 import { Button, Text, TextInput, View } from 'react-native';
 
-import ConditionalText from './ConditionalText';
-import ScreenView from './ScreenView';
+import ConditionalText from './components/ConditionalText';
+import ScreenView from './components/ScreenView';
 
 export const APP_NAME = 'AndroidStartupPerfApp';
